requirements.txt
# Python dependencies for Solana Agent Swarm

# Solana & Anchor
solana>=0.30.2
solders>=0.18.0
anchorpy>=0.18.0

# Async runtime
aiohttp>=3.9.0

# Data processing
numpy>=1.24.0
pandas>=2.0.0

# Configuration
python-dotenv>=1.0.0

# Logging
colorlog>=6.7.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Type hints
typing-extensions>=4.7.0

# Optional: AI/ML for learning agent
# scikit-learn>=1.3.0
# torch>=2.0.0
